+++ src/compas_timber/_fabrication/__init__.py
@@ -4,12 +4,18 @@ from .btlx_process import BTLxProcess
 from .btlx_process import OrientationType
 from .jack_cut import JackRafterCut
 from .jack_cut import JackRafterCutParams
+from .double_cut import DoubleCut
+from .double_cut import DoubleCutParams
 from .drilling import Drilling
 from .drilling import DrillingParams
 from .step_joint_notch import StepJointNotch
 from .step_joint_notch import StepJointNotchParams
 from .step_joint import StepJoint
 from .step_joint import StepJointParams
+from .dovetail_tenon import DovetailTenon
+from .dovetail_tenon import DovetailTenonParams
+from .dovetail_mortise import DovetailMortise
+from .dovetail_mortise import DovetailMortiseParams
 
 
 __all__ = [
@@ -17,10 +23,16 @@ __all__ = [
     "BTLxProcess",
     "OrientationType",
     "JackRafterCutParams",
+    "DoubleCut",
+    "DoubleCutParams",
     "Drilling",
     "DrillingParams",
     "StepJointNotch",
     "StepJointNotchParams",
     "StepJoint",
     "StepJointParams",
+    "DovetailTenon",
+    "DovetailTenonParams",
+    "DovetailMortise",
+    "DovetailMortiseParams",
 ]