+++ CHANGELOG.md
@@ -24,16 +24,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added new `compas_timber._fabrication.DrillingParams`.
 * Added new `compas_timber._fabrication.StepJoint`.
 * Added new `compas_timber._fabrication.StepJointNotch`.
+* Added new `compas_timber._fabrication.DovetailTenon`.
+* Added new `compas_timber._fabrication.DovetailMortise`.
 * Added new `compas_timber.connections.TStepJoint`.
+* Added new `compas_timber.connections.TDovetailJoint`.
 * Added new `utilities` module in `connections` package.
+* Added new `compas_timber._fabrication.DoubleCut`.
+* Added new `compas_timber.connections.TBirdsmouthJoint`.
 
 ### Changed
 
 * Changed incorrect import of `compas.geometry.intersection_line_plane()` to `compas_timber.utils.intersection_line_plane()`
+* Renamed `intersection_line_plane` to `intersection_line_plane_param`.
+* Renamed `intersection_line_line_3D` to `intersection_line_line_param`.
 * Added LayerName & ClearLayer Input to GH `BakeBoxMap` and `BakePlateBoxMap` Components.
 
 ### Removed
 
+* Removed module `compas_timber.utils.compas_extra`.
 
 ## [0.11.0] 2024-09-17